+++ ergo_std/ChangeLog.md
@@ -3,6 +3,9 @@
 ## Unreleased
 * Add `std:io:is-terminal` to check whether a stream is connected to a terminal.
 * Add `std:fs:file-type` to get the file type of a path.
+* Change `std:cache` to not cache errors by default.
+  * One can use the `allow-error` flag to cache errors if that behavior is
+    intended.
 
 ## 1.0.0-rc.1  -- 2021-12-08
 ### New Features